train/requirements.txt
torch==1.8.0
torchvision==0.9.0
opencv-python==4.3.0.38
psutil
h5py
matplotlib
scipy
scikit-image==0.16.2
tqdm
tensorboard
torchsummary
statistics
pyyaml
yacs
pathlib
termcolor
nvidia_ml_py3
pickle5
trimesh
numba
numpy-quaternion
seaborn
sympy
icecream
pytorch-lightning
shapely
vtk
tensorboardX
timm
iopath
